mejorando seguridad

+++ application/controllers/Ingreso.php
@@ -26,10 +26,10 @@ class Ingreso extends CI_Controller{
     private function acceso($id_rol){
         $rolusuario = $this->session_data['rol'];
         if($rolusuario[$id_rol-1]['rolusuario_asignado'] == 1){
-            return;
+            return true;
         }else{
             $data['_view'] = 'login/mensajeacceso';
-        $this->load->view('layouts/main',$data);
+            $this->load->view('layouts/main',$data);
         }
     }
     /*
@@ -37,23 +37,24 @@ class Ingreso extends CI_Controller{
      */
     function index()
     {
-        $this->acceso(16);
-        //$data['ingreso'] = $this->Ingreso_model->get_all_ingreso();
-        $this->load->model('Institucion_model');
-        $data['institucion'] = $this->Institucion_model->get_all_institucion();
-        $this->load->model('Unidad_model');
-        $data['all_unidad'] = $this->Unidad_model->get_all_unidad();
-        $this->load->model('Programa_model');
-        $data['all_programa'] = $this->Programa_model->get_all_programa();
-        $tipo = 1;
-        $this->load->model('Estado_model');
-        $data['all_estado'] = $this->Estado_model->get_estado_tipo($tipo);
-        $data['_view'] = 'ingreso/index';
-        $this->load->view('layouts/main',$data);
+        if($this->acceso(26)){
+            $data['rolusuario'] = $this->session_data['rol'];
+            //$data['ingreso'] = $this->Ingreso_model->get_all_ingreso();
+            $this->load->model('Institucion_model');
+            $data['institucion'] = $this->Institucion_model->get_all_institucion();
+            $this->load->model('Unidad_model');
+            $data['all_unidad'] = $this->Unidad_model->get_all_unidad();
+            $this->load->model('Programa_model');
+            $data['all_programa'] = $this->Programa_model->get_all_programa();
+            $tipo = 1;
+            $this->load->model('Estado_model');
+            $data['all_estado'] = $this->Estado_model->get_estado_tipo($tipo);
+            $data['_view'] = 'ingreso/index';
+            $this->load->view('layouts/main',$data);
+        }
     }
     function buscar_ingresoexcel()
      {
-        $this->acceso(16);
         if ($this->input->is_ajax_request())
         {
             $parametro = $this->input->post('parametro');
@@ -68,7 +69,6 @@ class Ingreso extends CI_Controller{
     }
      function buscar50ingreso()
     {
-         $this->acceso(16);
         if ($this->input->is_ajax_request())
         {
           
@@ -82,7 +82,6 @@ class Ingreso extends CI_Controller{
     }
      function buscarallingreso()
     {
-         $this->acceso(16);
         if ($this->input->is_ajax_request())
         {
           
@@ -96,7 +95,6 @@ class Ingreso extends CI_Controller{
     }
      function buscarporingreso()
     {
-         $this->acceso(16);
         if ($this->input->is_ajax_request())
         {
             $parametro = $this->input->post('parametro');
@@ -112,9 +110,8 @@ class Ingreso extends CI_Controller{
 
     function crear()
     {
-        $this->acceso(16);
-        $usuario_id = 1;
-        $gestion_id = 1;
+        $usuario_id = $this->session_data['usuario_id'];
+        $gestion_id = $this->session_data['gestion_id'];
         $numrec = $this->Ingreso_model->get_numero();
         $numero = $numrec['numero'];
         $ingreso_id = $this->Ingreso_model->crear_ingreso($usuario_id,$gestion_id,$numero);
@@ -124,9 +121,8 @@ class Ingreso extends CI_Controller{
     }
     function nuevo()
     {
-        $this->acceso(16);
-        $usuario_id = 1;
-        $gestion_id = 1; 
+        $usuario_id = $this->session_data['usuario_id'];
+        $gestion_id = $this->session_data['gestion_id']; 
         $ingreso_id = $this->Ingreso_model->crear_ingreso_extra($usuario_id,$gestion_id);
          
         redirect('ingreso/edit/'.$ingreso_id);
@@ -136,33 +132,33 @@ class Ingreso extends CI_Controller{
      */
     function add($ingreso_id)
     {
-        $this->acceso(16);
-        $data['ingreso_id'] = $ingreso_id;
-        $data['ingreso'] = $this->Ingreso_model->get_ing_completo($ingreso_id);
-        $this->load->model('Programa_model');
-        $data['all_programa'] = $this->Programa_model->get_all_programa();
-        $this->load->model('Responsable_model');
-        $data['responsable'] = $this->Responsable_model->get_all_responsable();
-        $data['pedidos'] = $this->Ingreso_model->get_pedidos($ingreso_id);
-        $data['facturas'] = $this->Ingreso_model->get_facturas($ingreso_id);
-        $data['numero'] = $this->Ingreso_model->get_numero();
+        if($this->acceso(25)){
+            $data['ingreso_id'] = $ingreso_id;
+            $data['ingreso'] = $this->Ingreso_model->get_ing_completo($ingreso_id);
+            $this->load->model('Programa_model');
+            $data['all_programa'] = $this->Programa_model->get_all_programa();
+            $this->load->model('Responsable_model');
+            $data['responsable'] = $this->Responsable_model->get_all_responsable();
+            $data['pedidos'] = $this->Ingreso_model->get_pedidos($ingreso_id);
+            $data['facturas'] = $this->Ingreso_model->get_facturas($ingreso_id);
+            $data['numero'] = $this->Ingreso_model->get_numero();
 
-        $this->load->model('Proveedor_model');
-        $data['proveedor'] = $this->Proveedor_model->get_all_proveedor();
+            $this->load->model('Proveedor_model');
+            $data['proveedor'] = $this->Proveedor_model->get_all_proveedor();
 
-        $this->load->model('Pedido_model');
-        $data['all_pedido'] = $this->Ingreso_model->get_pedido_pendiente();
+            $this->load->model('Pedido_model');
+            $data['all_pedido'] = $this->Ingreso_model->get_pedido_pendiente();
 
-        $this->load->model('Usuario_model');
-        $data['all_usuario'] = $this->Usuario_model->get_all_usuario();
+            $this->load->model('Usuario_model');
+            $data['all_usuario'] = $this->Usuario_model->get_all_usuario();
 
-        $data['_view'] = 'ingreso/add';
-        $this->load->view('layouts/main',$data);
+            $data['_view'] = 'ingreso/add';
+            $this->load->view('layouts/main',$data);
+        }
     }
 
     function responsables()
     {
-        $this->acceso(16);
         if ($this->input->is_ajax_request()) {
             $datos = $this->Responsable_model->get_all_responsable();
             echo json_encode($datos);
@@ -176,7 +172,6 @@ class Ingreso extends CI_Controller{
 
     function buscaringreso()
     {
-        $this->acceso(16);
         if ($this->input->is_ajax_request()) {
             $parametro = $this->input->post('parametro');   
             if ($parametro!=""){
@@ -193,7 +188,6 @@ class Ingreso extends CI_Controller{
 
 function detalleingreso()
 {
-    $this->acceso(16);
     if ($this->input->is_ajax_request()) {  
     $ingreso_id = $this->input->post('ingreso_id');
     $datos = $this->Ingreso_model->get_detalle_ingreso_aux($ingreso_id);
@@ -210,7 +204,6 @@ else
 
 function pedidosunidad()
 {
-    $this->acceso(16);
    if ($this->input->is_ajax_request()) {  
     $unidad_id = $this->input->post('unidad_id');
     $datos = $this->Ingreso_model->get_pedidounidad($unidad_id);
@@ -226,7 +219,6 @@ else
 }
 function pedidosfiltro()
 {
-    $this->acceso(16);
    if ($this->input->is_ajax_request()) {  
     $filtro = $this->input->post('filtro');
     $datos = $this->Ingreso_model->get_pedidofiltro($filtro);
@@ -243,7 +235,6 @@ else
 
 function ingresararticulo()
 {
-    $this->acceso(16);
     if ($this->input->is_ajax_request()) {
         $ingreso_id = $this->input->post('ingreso_id');
         $articulo_id = $this->input->post('articulo_id');
@@ -270,25 +261,24 @@ function ingresararticulo()
         ".$cantidad.",
         ".$factura_numero."
         
-    )";
+        )";
 
-    $this->db->query($sql);
-    $detalles = $this->db->insert_id();
-  
-    $datos = $this->Ingreso_model->get_detalle_ingreso_aux($ingreso_id);
-    if(isset($datos)){
-        echo json_encode($datos);
-    }else echo json_encode(null);
-}
-else
-{                 
-    show_404();
-}          
+        $this->db->query($sql);
+        $detalles = $this->db->insert_id();
+
+        $datos = $this->Ingreso_model->get_detalle_ingreso_aux($ingreso_id);
+        if(isset($datos)){
+            echo json_encode($datos);
+        }else echo json_encode(null);
+    }
+    else
+    {                 
+        show_404();
+    }          
 }
 
 function updateDetalle()
 {
-    $this->acceso(16);
     $detalleing_id = $this->input->post('detalleing_id');
     $cantidad = $this->input->post('cantidad'); 
     $precio = $this->input->post('precio');   
@@ -312,36 +302,30 @@ function updateDetalle()
 }
 function quitar($detalleing_id)
 {
-     $this->acceso(16);
- $sql = "delete from detalle_ingreso_aux where detalleing_id = ".$detalleing_id;
- $this->db->query($sql);
- 
- return true;
- 
+    $sql = "delete from detalle_ingreso_aux where detalleing_id = ".$detalleing_id;
+    $this->db->query($sql);
+
+    return true;
 }
 function quitarpedido()
 {
-    $this->acceso(16);
- if ($this->input->is_ajax_request()) { 
-   $pedido_id = $this->input->post('pedido_id');  
-   $ingreso_id = $this->input->post('ingreso_id');     
- $sql = "update pedido set ingreso_id=0, estado_id=6 where pedido_id = ".$pedido_id;
- $this->db->query($sql);
-  $datos =  $this->Ingreso_model->get_pedidos($ingreso_id);
-        
+    if($this->input->is_ajax_request()) {
+        $pedido_id = $this->input->post('pedido_id');  
+        $ingreso_id = $this->input->post('ingreso_id');     
+        $sql = "update pedido set ingreso_id=0, estado_id=6 where pedido_id = ".$pedido_id;
+        $this->db->query($sql);
+        $datos =  $this->Ingreso_model->get_pedidos($ingreso_id);
         if(isset($datos)){
-                        echo json_encode($datos);
-                    }else echo json_encode(null);
-    }
-        else
-        {                 
-                    show_404();
-        }          
+            echo json_encode($datos);
+        }else echo json_encode(null);
+    }else
+    {
+                show_404();
+    }          
     }
 
 function quitarfactura()
 {
-    $this->acceso(16);
    if ($this->input->is_ajax_request()) { 
    $factura_id = $this->input->post('factura_id');  
    $ingreso_id = $this->input->post('ingreso_id');  
@@ -358,9 +342,8 @@ function quitarfactura()
         }          
     }
 
-function ingresoapedido()
+    function ingresoapedido()
     {
-    $this->acceso(16);
          if ($this->input->is_ajax_request()) {
    
         $pedido_id = $this->input->post('pedido_id');
@@ -375,18 +358,15 @@ function ingresoapedido()
         if(isset($datos)){
                         echo json_encode($datos);
                     }else echo json_encode(null);
-    }
-        else
-        {                 
-                    show_404();
-        }          
+        }else{                 
+            show_404();
+        }
     }
 
-function crearfactura()
+    function crearfactura()
     {
-    $this->acceso(16);
-        $usuario_id = 1;
-        $gestion_id = 1;
+        $usuario_id = $this->session_data['usuario_id'];
+        $gestion_id = $this->session_data['gestion_id'];
         $estado_id = 1;
          if ($this->input->is_ajax_request()) {
    
@@ -456,41 +436,29 @@ function crearfactura()
         }          
     }
 
-function cambiarproveedor()
+    function cambiarproveedor()
     {
-    $this->acceso(16);
+        if ($this->input->is_ajax_request()) {
+            $proveedor_id = $this->input->post('proveedor_id');
+            $ingreso_id = $this->input->post('ingreso_id');
 
-         if ($this->input->is_ajax_request()) {
-   
-        $proveedor_id = $this->input->post('proveedor_id');
-        $ingreso_id = $this->input->post('ingreso_id');
-       
-        
-        $this->load->model('ingreso_model');
-  
-        $this->Ingreso_model->cambiar_proveedor($ingreso_id,$proveedor_id);
-       
-        $datos =  $this->Ingreso_model->get_ingreso_proveedor($ingreso_id,$proveedor_id);
-        
+            $this->load->model('ingreso_model');
+            $this->Ingreso_model->cambiar_proveedor($ingreso_id,$proveedor_id);
 
-        if(isset($datos)){
-                        echo json_encode($datos);
-                    }else echo json_encode(null);
-    }
-        else
-        {                 
-                    show_404();
+            $datos =  $this->Ingreso_model->get_ingreso_proveedor($ingreso_id,$proveedor_id);
+            if(isset($datos)){
+                echo json_encode($datos);
+            }else echo json_encode(null);
+        }else{
+            show_404();
         }          
     }
 function cambiarpedido()
     {
-    $this->acceso(16);
-
-         if ($this->input->is_ajax_request()) {
+        if ($this->input->is_ajax_request()) {
    
         $pedido_id = $this->input->post('pedido_id');
         $ingreso_id = $this->input->post('ingreso_id');
-       
         
         $this->load->model('ingreso_model');
   
@@ -501,18 +469,15 @@ function cambiarpedido()
         if(isset($datos)){
                         echo json_encode($datos);
                     }else echo json_encode(null);
-    }
-        else
-        {                 
-                    show_404();
-        }          
+    }else{                 
+        show_404();
+    }          
     }
 
 function finalizaringreso($ingreso_id)
 {
-    $this->acceso(16);
-    $usuario_id = 1;
-    $gestion_id = 1;
+    $usuario_id = $this->session_data['usuario_id'];
+    $gestion_id = $this->session_data['gestion_id'];
     $estado_id = 1;
     $programa_id = $this->input->post('programa_id');
     $proveedor_id = $this->input->post('proveedor_id');
@@ -570,14 +535,12 @@ $this->db->query($pedidos);
 
  $eliminar_aux = "DELETE FROM detalle_ingreso_aux WHERE ingreso_id=".$ingreso_id." ";
    $this->db->query($eliminar_aux);
- 
 }
 
 function actualizarzaringreso($ingreso_id)
 {
-    $this->acceso(16);
- $usuario_id = 1;
- $gestion_id = 1;
+    $usuario_id = $this->session_data['usuario_id'];
+    $gestion_id = $this->session_data['gestion_id'];
  $estado_id = 1;
  $programa_id = $this->input->post('programa_id');
  //$proveedor_id = $this->input->post('proveedor_id');
@@ -659,7 +622,7 @@ $num_actual = $this->db->query($numero_actual)->result_array();
      */
     function editar($ingreso_id)
     {
-        $this->acceso(16);
+        $this->acceso(30);
 
             ///////////1.  BORRAR AUX DE LA COMPRA//////////
     $eliminar_aux = "DELETE FROM detalle_ingreso_aux WHERE ingreso_id=".$ingreso_id." ";
@@ -696,7 +659,7 @@ $num_actual = $this->db->query($numero_actual)->result_array();
     }
     function edit($ingreso_id)
     {
-        $this->acceso(16);
+        if($this->acceso(30)){
             $data['ingreso_id'] = $ingreso_id;
             $data['ingreso'] = $this->Ingreso_model->get_ing_mascompleto($ingreso_id);
             $data['pedidos'] = $this->Ingreso_model->get_pedidos($ingreso_id);
@@ -716,14 +679,15 @@ $num_actual = $this->db->query($numero_actual)->result_array();
             
             $data['_view'] = 'ingreso/edit';
             $this->load->view('layouts/main',$data);
+        }
 				
     }
 
     function pdf($ingreso_id)
     {
-        $this->acceso(16);
+        if($this->acceso(31)){
         // check if the ingreso exists before trying to edit it
-            $gestion_id =1;
+            $gestion_id = $this->session_data['gestion_id'];
             $data['ingreso_id'] = $ingreso_id;
             $data['datos'] = $this->Ingreso_model->get_ing_mascompleto($ingreso_id);
             $data['detalle_ingreso'] = $this->Ingreso_model->get_detalle_ingreso($ingreso_id);
@@ -737,7 +701,7 @@ $num_actual = $this->db->query($numero_actual)->result_array();
 
             $data['_view'] = 'ingreso/pdf';
             $this->load->view('layouts/main',$data);
-                
+        }
     }
     
     /*
@@ -745,37 +709,36 @@ $num_actual = $this->db->query($numero_actual)->result_array();
      */
     function remove($ingreso_id)
     {
-        $this->acceso(17);
-        $ingreso = $this->Ingreso_model->get_ingreso($ingreso_id);
+        if($this->acceso(17)){
+            $ingreso = $this->Ingreso_model->get_ingreso($ingreso_id);
 
-        // check if the ingreso exists before trying to delete it
-        if(isset($ingreso['ingreso_id']))
-        {
-            $this->Ingreso_model->delete_ingreso($ingreso_id);
-            redirect('ingreso/index');
+            // check if the ingreso exists before trying to delete it
+            if(isset($ingreso['ingreso_id']))
+            {
+                $this->Ingreso_model->delete_ingreso($ingreso_id);
+                redirect('ingreso/index');
+            }
+            else
+                show_error('The ingreso you are trying to delete does not exist.');
         }
-        else
-            show_error('The ingreso you are trying to delete does not exist.');
     }
     function eliminar()
     {
-        $this->acceso(16);
-        
-        $ingreso_id = $this->input->post('ingreso_id'); 
-        $ingreso = $this->Ingreso_model->get_ingreso($ingreso_id);
-        // check if the programa exists before trying to delete it
-        
-$pedi = "update pedido  SET estado_id=6 where ingreso_id = ".$ingreso_id." ";
- $this->db->query($pedi);
- $pqs = "delete from detalle_ingreso where ingreso_id = ".$ingreso_id." ";
- $this->db->query($pqs);
-  $ptq = "delete from factura where ingreso_id = ".$ingreso_id." ";
- $this->db->query($ptq);
-$sql = "delete from ingreso where ingreso_id = ".$ingreso_id." ";
- $this->db->query($sql);
-
-       return true;
-    } 
+        if($this->acceso(17)){
+            $ingreso_id = $this->input->post('ingreso_id'); 
+            $ingreso = $this->Ingreso_model->get_ingreso($ingreso_id);
+            // check if the programa exists before trying to delete it
+            $pedi = "update pedido  SET estado_id=6 where ingreso_id = ".$ingreso_id." ";
+            $this->db->query($pedi);
+            $pqs = "delete from detalle_ingreso where ingreso_id = ".$ingreso_id." ";
+            $this->db->query($pqs);
+            $ptq = "delete from factura where ingreso_id = ".$ingreso_id." ";
+            $this->db->query($ptq);
+            $sql = "delete from ingreso where ingreso_id = ".$ingreso_id." ";
+            $this->db->query($sql);
+         return true;
+        }
+    }
 }
